revision [12:32pm]
loop to fix the fact that:

when you press d to delete
goto beginning;

it deleted the list but then still asked the user to search the sentence, of which there would be none. 

head->length;                // the loop performs a conditional comparison for length != '0'
If theres length ---> 
 reveal the initial search functionality and break upon its completion which would be the returning of the characters that were met by the search param

if theres no length. invoke the printf() housed by the else block. immediately break.
the program proceeds past the search, never asking the user to search for a character if there is no data populated in the words struct (which would be linked list)